+++ tests/mock_hass/homeassistant/helpers/restore_state.py
@@ -1,0 +1,5 @@
+"""Mock restore_state for homeassistant.helpers."""
+
+class RestoreEntity:
+    async def async_get_last_state(self):
+        return None