+++ tests/mock_hass/pytest_homeassistant_custom_component/common.py
@@ -1,9 +1,13 @@
 class MockConfigEntry:
-    def __init__(self, domain, unique_id, data, entry_id="test"):
+    def __init__(self, domain="hcu_integration", unique_id="test", data=None, entry_id="test", options=None, title=None, **kwargs):
         self.domain = domain
         self.unique_id = unique_id
-        self.data = data
+        self.data = data or {}
         self.entry_id = entry_id
+        self.options = options or {}
+        self.title = title
+        for k, v in kwargs.items():
+            setattr(self, k, v)
         
     def add_to_hass(self, hass):
         pass